requirements.txt
absl-py==2.1.0
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.3.1
cycler==0.12.1
filelock==3.17.0
flatbuffers==25.1.24
fonttools==4.56.0
fsspec==2025.2.0
ftfy==6.3.1
idna==3.10
jax==0.5.0
jaxlib==0.5.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==4.25.6
psutil==6.1.1
py-cpuinfo==9.0.0
pycparser==2.22
pyparsing==3.2.1
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scipy==1.15.1
seaborn==0.13.2
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
sympy==1.13.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.73
ultralytics-thop==2.0.14
urllib3==2.3.0
wcwidth==0.2.13
